+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+# Changelog
+
+All notable changes to `tester-phpunit` will be documented in this file
+
+## 0.1.2
+- 分析五種狀態 passed, failed, error, skipped, incomplete
+- PHPUnitReporter 更名為 JUnitParser
+
 ## 0.1.0 - First Release
-* Just block the name in the atom registry